Remove old test, rename new test
